refactor(CreateListButton): dialog out of search route to component
